requirements.txt
udsoncan
can-isotp
sa2-seed-key
tqdm
pycryptodome
wxpython==4.1.1
python-can
pyyaml
pyserial
nicegui